test(orderbook): split benchmarks by subject
