+++ exchanges/orderbook/levels_test.go
@@ -99,15 +99,6 @@ func TestLoad(t *testing.T) {
 	Check(t, list, 0, 0, 0)
 }
 
-// 27906781	        42.4 ns/op	       0 B/op	       0 allocs/op (old)
-// 84119028	        13.87 ns/op	       0 B/op	       0 allocs/op (new)
-func BenchmarkLoad(b *testing.B) {
-	ts := Levels{}
-	for b.Loop() {
-		ts.load(ask)
-	}
-}
-
 func TestUpdateInsertByPrice(t *testing.T) {
 	a := askLevels{}
 	asksSnapshot := Levels{
@@ -244,51 +235,6 @@ func TestUpdateInsertByPrice(t *testing.T) {
 	Check(t, b, 6, 36, 6)
 }
 
-// 134830672	         9.83 ns/op	       0 B/op	       0 allocs/op (old)
-// 206689897	         5.761 ns/op	   0 B/op	       0 allocs/op (new)
-func BenchmarkUpdateInsertByPrice_Amend(b *testing.B) {
-	a := askLevels{}
-	a.load(ask)
-
-	updates := Levels{
-		{
-			Price:  1337, // Amend
-			Amount: 2,
-		},
-		{
-			Price:  1337, // Amend
-			Amount: 1,
-		},
-	}
-
-	for b.Loop() {
-		a.updateInsertByPrice(updates, 0)
-	}
-}
-
-// 49763002	        24.9 ns/op	       0 B/op	       0 allocs/op (old)
-// 25662849	        45.32 ns/op	       0 B/op	       0 allocs/op (new)
-func BenchmarkUpdateInsertByPrice_Insert_Delete(b *testing.B) {
-	a := askLevels{}
-
-	a.load(ask)
-
-	updates := Levels{
-		{
-			Price:  1337.5, // Insert
-			Amount: 2,
-		},
-		{
-			Price:  1337.5, // Delete
-			Amount: 0,
-		},
-	}
-
-	for b.Loop() {
-		a.updateInsertByPrice(updates, 0)
-	}
-}
-
 func TestUpdateByID(t *testing.T) {
 	a := askLevels{}
 	asksSnapshot := Levels{
@@ -338,28 +284,6 @@ func TestUpdateByID(t *testing.T) {
 	}
 }
 
-// 46043871	        25.9 ns/op	       0 B/op	       0 allocs/op (old)
-// 63445401	        18.51 ns/op	       0 B/op	       0 allocs/op (new)
-func BenchmarkUpdateByID(b *testing.B) {
-	asks := Levels{}
-	asksSnapshot := Levels{
-		{Price: 1, Amount: 1, ID: 1},
-		{Price: 3, Amount: 1, ID: 3},
-		{Price: 5, Amount: 1, ID: 5},
-		{Price: 7, Amount: 1, ID: 7},
-		{Price: 9, Amount: 1, ID: 9},
-		{Price: 11, Amount: 1, ID: 11},
-	}
-	asks.load(asksSnapshot)
-
-	for b.Loop() {
-		err := asks.updateByID(asksSnapshot)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
 func TestDeleteByID(t *testing.T) {
 	a := askLevels{}
 	asksSnapshot := Levels{
@@ -404,28 +328,6 @@ func TestDeleteByID(t *testing.T) {
 	err = a.deleteByID(Levels{{Price: 11, Amount: 1, ID: 1337}}, true)
 	if err != nil {
 		t.Fatal(err)
-	}
-}
-
-// 26724331	        44.69 ns/op	       0 B/op	       0 allocs/op
-func BenchmarkDeleteByID(b *testing.B) {
-	asks := Levels{}
-	asksSnapshot := Levels{
-		{Price: 1, Amount: 1, ID: 1},
-		{Price: 3, Amount: 1, ID: 3},
-		{Price: 5, Amount: 1, ID: 5},
-		{Price: 7, Amount: 1, ID: 7},
-		{Price: 9, Amount: 1, ID: 9},
-		{Price: 11, Amount: 1, ID: 11},
-	}
-	asks.load(asksSnapshot)
-
-	for b.Loop() {
-		err := asks.deleteByID(asksSnapshot, false)
-		if err != nil {
-			b.Fatal(err)
-		}
-		asks.load(asksSnapshot) // reset
 	}
 }
 
@@ -682,27 +584,6 @@ func TestUpdateInsertByIDAsk(t *testing.T) {
 	Check(t, a, 14, 87, 7)
 }
 
-// 21614455	        81.74 ns/op	       0 B/op	       0 allocs/op
-func BenchmarkUpdateInsertByID_asks(b *testing.B) {
-	asks := Levels{}
-	asksSnapshot := Levels{
-		{Price: 1, Amount: 1, ID: 1},
-		{Price: 3, Amount: 1, ID: 3},
-		{Price: 5, Amount: 1, ID: 5},
-		{Price: 7, Amount: 1, ID: 7},
-		{Price: 9, Amount: 1, ID: 9},
-		{Price: 11, Amount: 1, ID: 11},
-	}
-	asks.load(asksSnapshot)
-
-	for b.Loop() {
-		err := asks.updateInsertByID(asksSnapshot, askCompare)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
 func TestUpdateInsertByIDBids(t *testing.T) {
 	b := bidLevels{}
 	bidsSnapshot := Levels{
@@ -951,28 +832,6 @@ func TestUpdateInsertByIDBids(t *testing.T) {
 	}
 
 	Check(t, b, 14, 87, 7)
-}
-
-// 20328886	        59.94 ns/op	       0 B/op	       0 allocs/op
-func BenchmarkUpdateInsertByID_bids(b *testing.B) {
-	bids := Levels{}
-	bidsSnapshot := Levels{
-		{Price: 0.5, Amount: 2, ID: 0},
-		{Price: 1, Amount: 2, ID: 1},
-		{Price: 3, Amount: 2, ID: 3},
-		{Price: 12, Amount: 2, ID: 5},
-		{Price: 7, Amount: 2, ID: 7},
-		{Price: 9, Amount: 2, ID: 9},
-		{Price: 11, Amount: 2, ID: 11},
-	}
-	bids.load(bidsSnapshot)
-
-	for b.Loop() {
-		err := bids.updateInsertByID(bidsSnapshot, bidCompare)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
 }
 
 func TestInsertUpdatesBid(t *testing.T) {
@@ -1894,22 +1753,4 @@ func TestFinalizeFields(t *testing.T) {
 	mov, err := m.finalizeFields(20000*151.11585, 20000, 151.08, 0, false)
 	assert.NoError(t, err, "finalizeFields should not error")
 	assert.InDelta(t, 717.0, mov.SlippageCost, 0.000000001, "SlippageCost should be correct")
-}
-
-// 8384302	       150.9 ns/op	     480 B/op	       1 allocs/op
-func BenchmarkRetrieve(b *testing.B) {
-	asks := Levels{}
-	asksSnapshot := Levels{
-		{Price: 1, Amount: 1, ID: 1},
-		{Price: 3, Amount: 1, ID: 3},
-		{Price: 5, Amount: 1, ID: 5},
-		{Price: 7, Amount: 1, ID: 7},
-		{Price: 9, Amount: 1, ID: 9},
-		{Price: 11, Amount: 1, ID: 11},
-	}
-	asks.load(asksSnapshot)
-
-	for b.Loop() {
-		_ = asks.retrieve(6)
-	}
 }
